+++ Sources/IOSScene/WordChecking/Models/MemorizingCount.swift
@@ -1,0 +1,14 @@
+//
+//  MemorizingCount.swift
+//  IOSScene_WordChecking
+//
+//  Created by Jaewon Yun on 5/29/24.
+//  Copyright © 2024 woin2ee. All rights reserved.
+//
+
+import Foundation
+
+struct MemorizingCount {
+    var checked: Int
+    var total: Int
+}